+++ src/main/kotlin/us/frollo/frollosdk/database/SDKDatabase.kt
@@ -520,6 +520,7 @@ abstract class SDKDatabase : RoomDatabase() {
                 // 4) New table - addresses
                 // 5) Alter account table - add new column payids
                 // 6) Alter transaction_model table - add new column reference, reason
+                // 7) Alter cards table - add digital wallet column
 
                 database.execSQL("CREATE TABLE IF NOT EXISTS `payday` (`status` TEXT NOT NULL, `frequency` TEXT NOT NULL, `next_date` TEXT, `previous_date` TEXT, `payday_id` INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL)")
 
@@ -543,6 +544,7 @@ abstract class SDKDatabase : RoomDatabase() {
 
                 database.execSQL("ALTER TABLE `transaction_model` ADD COLUMN `reference` TEXT")
                 database.execSQL("ALTER TABLE `transaction_model` ADD COLUMN `reason` TEXT")
+                database.execSQL("ALTER TABLE `card` ADD COLUMN `digital_wallets` TEXT")
             }
         }
     }