+++ src/main/kotlin/us/frollo/frollosdk/mapping/CardsMapping.kt
@@ -34,5 +34,6 @@ internal fun CardResponse.toCard(): Card =
         cardholderName = cardholderName,
         type = type,
         issuer = issuer,
-        pinSetDate = pinSetDate
+        pinSetDate = pinSetDate,
+        digitalWallets = digitalWallets
     )